+++ NEWS.md
@@ -13,12 +13,16 @@ Fixes:
   * Fix estimation of passed fish when using intra-array estimates.
   * Fix split CJS's not receiving information about intra-array estimates.
 
+Changes:
+  * `array.overview` is now called `group.overview`.
+
 Enhancements:
   * The object `detections` now contains a "Valid" column, similarly to the `movements` objects.
   * Invalid detections are now painted in grey in the individual detection plots.
   * Check for duplicated detections and give the user a chance to clear them out before continuing.
   * Prevent unhandled crashes if the 'replicates' argument is badly structured.
   * Show number of released fish per release location in the release sites table.
+  * Detailed progression per group and release site is now exported through a `release.overview` object.
 
 
 ## actel 0.1.2